requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic[email]==2.7.1
scikit-learn==1.4.2
joblib==1.4.2
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
alembic==1.13.1
groq==0.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
